# internal/api/http/openapi.yaml
openapi: 3.1.0
info:
  title: Plowered API
  version: "1.0.0"
  description: |
    Plowered is a self-hosted, AI-native data context platform: catalog,
    orchestration, quality, governance, lineage, ETL/ELT, MCP, all in one.

    All endpoints sit behind the same auth chain — session cookie (web UI)
    or Bearer JWT (integrations). Tenant isolation is enforced server-side
    from the authenticated principal; you never pass tenant_id in payloads
    or query strings.

    Identifiers throughout are UUIDs unless documented otherwise.
  contact:
    name: Plowered
    url: https://github.com/Satyaamm/plowered
servers:
  - url: http://localhost:8080
    description: Local dev
security:
  - cookieAuth: []
  - bearerAuth: []
...
paths:
...
  /v1/assets:
    get:
      tags: [catalog]
      summary: List assets (paginated)
      parameters:
        - in: query
          name: type
          schema: { type: string, enum: [database, schema, table, view, column, dashboard, report, dbt_model, dag, ml_model, glossary_term] }
        - in: query
          name: page_size
          schema: { type: integer, default: 50, maximum: 500 }
        - in: query
          name: page_token
          schema: { type: string }
      responses:
        "200":
          description: Asset page
          content:
            application/json:
              schema:
                type: object
                properties:
                  assets:
                    type: array
                    items: { $ref: "#/components/schemas/Asset" }
                  next_page_token: { type: string }
    post:
      tags: [catalog]
      summary: Upsert an asset by qualified_name
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/Asset" } } }
      responses:
        "201": { description: Asset created or replaced }
  /v1/assets/{id}:
    get:
      tags: [catalog]
      summary: Get an asset by UUID
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200": { description: Asset, content: { application/json: { schema: { $ref: "#/components/schemas/Asset" } } } }
        "404": { description: Not found }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: plowered_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Asset:
      type: object
      properties:
        id:             { type: string, format: uuid }
        tenant_id:      { type: string, format: uuid }
        qualified_name: { type: string }
        type:           { type: string }
        name:           { type: string }
        description:    { type: string }
        description_ai: { type: string }
        trust:          { type: string, enum: [unverified, draft, reviewed, certified, deprecated] }
        tags:           { type: array, items: { type: string } }
        owners:         { type: array, items: { type: string } }
        properties:     { type: object, additionalProperties: true }
        created_at:     { type: string, format: date-time }
        updated_at:     { type: string, format: date-time }
...